+++ GuiModule.h
@@ -8,7 +8,7 @@ public:
     static void printBoard(const Board &chessboard);
     static void printWelcomeInfo();
 private:
-    static std::string numToSign(int8 a, int8 color);
+    static std::string numToSign(int a, int color);
 };
